+++ limited_area/mpas_to_mpas/RemapperEdge.cpp
@@ -76,9 +76,9 @@ void RemapperEdge::computeWeightsEdge(int nCellsSrc, int nEdgesDst, int nVertLev
 	nHDstPts = nEdgesDst;
 	maxHSrcPts = maxEdges;
 	nHSrcPts = new int[nHDstPts];   // set to all 1 for now...
-	HSrcPts = new int[nHDstPts * maxHSrcPts];
+	HSrcPts = new int[(size_t)nHDstPts * (size_t)maxHSrcPts];
 	HSrcPts2d = allocate_2d<int>(nHDstPts, maxHSrcPts, HSrcPts);
-	HSrcWghts = new float[nHDstPts * maxHSrcPts];
+	HSrcWghts = new float[(size_t)nHDstPts * (size_t)maxHSrcPts];
 	HSrcWghts2d = allocate_2d<float>(nHDstPts, maxHSrcPts, HSrcWghts);
 	HDstMask = new unsigned char[nHDstPts];
 
@@ -108,11 +108,11 @@ void RemapperEdge::computeWeightsEdge(int nCellsSrc, int nEdgesDst, int nVertLev
 		nVDstPts = nVertLevelsDst;
 		nVSrcLevels = nVertLevelsSrc;
 		maxVSrcPts = 2;
-		nVSrcPts = new int[nHDstPts * nVDstPts];   // set to all 1 for now...
+		nVSrcPts = new int[(size_t)nHDstPts * (size_t)nVDstPts];   // set to all 1 for now...
 		nVSrcPts2d = allocate_2d<int>(nHDstPts, nVDstPts, nVSrcPts);
-		VSrcPts = new int[nHDstPts * nVDstPts * maxVSrcPts];
+		VSrcPts = new int[(size_t)nHDstPts * (size_t)nVDstPts * (size_t)maxVSrcPts];
 		VSrcPts3d = allocate_3d<int>(nHDstPts, nVDstPts, maxVSrcPts, VSrcPts);
-		VSrcWghts = new float[nHDstPts * nVDstPts * maxVSrcPts];
+		VSrcWghts = new float[(size_t)nHDstPts * (size_t)nVDstPts * (size_t)maxVSrcPts];
 		VSrcWghts3d = allocate_3d<float>(nHDstPts, nVDstPts, maxVSrcPts, VSrcWghts);
 
 #pragma omp parallel for private(tempLevels)
